Update config
JP updated the file names for 1) the new versions of IL and IN data and 2) the UMW data which are now csv files for each individual UMW state.  The UMW files are in both a zip file and a csv because the csv files were too large to directly upload to the wiki.